+++ src/WatchCompass.Application/Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.DependencyInjection;
 using WatchCompass.Application.UseCases.Recommendations;
 using WatchCompass.Application.UseCases.MovieDetails;
 using WatchCompass.Application.UseCases.SearchMovies;
+using WatchCompass.Application.UseCases.Genres;
 
 namespace WatchCompass.Application.DependencyInjection;
 
@@ -12,6 +13,7 @@ public static class ApplicationServiceCollectionExtensions
         services.AddScoped<SearchMoviesUseCase>();
         services.AddScoped<GetRecommendationsUseCase>();
         services.AddScoped<GetMovieDetailsUseCase>();
+        services.AddScoped<GetGenresUseCase>();
         return services;
     }
 }